refactor: 스피너 동작 방식 변경
